requirements.txt
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.120.0
icalendar==5.0.11
streamlit==1.32.0
python-dotenv==1.0.1
requests==2.31.0 